+++ data/smolLM_fineweb-edu/prepare.py
@@ -74,7 +74,7 @@ if __name__ == "__main__":
         ids = enc.encode(text)
         ids.append(enc.eot_token)
 
-        arr = np.array(ids, dtype=np.uint32)
+        arr = np.array(ids, dtype=np.uint16)
 
         # split + write
         if is_val(text):
