+++ data/smolLM_python-edu/prepare.py
@@ -72,7 +72,7 @@ for example in tqdm(dataset, desc="Streaming + tokenizing"):
     ids = enc.encode(text)
     ids.append(enc.eot_token)
 
-    arr = np.array(ids, dtype=np.uint32)
+    arr = np.array(ids, dtype=np.uint16)
 
     # split
     if is_val(text):
